moved species logic from index.js to new structure under userController.js
